Remove unused LSP misc files workspace and refactor tests
