+++ src/LanguageServer/Microsoft.CodeAnalysis.LanguageServer.UnitTests/MiscellaneousFiles/LspMiscellaneousFilesWorkspaceTests.cs
@@ -2,30 +2,27 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using System.Composition;
-using System.IO;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Features.Workspaces;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.LanguageServer.UnitTests.Miscellaneous;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Roslyn.LanguageServer.Protocol;
 using Roslyn.Test.Utilities;
-using Xunit;
 using Xunit.Abstractions;
 using LSP = Roslyn.LanguageServer.Protocol;
 
-namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Miscellaneous;
+namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.MiscellaneousFiles;
 
-public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLanguageServerProtocolTests
+/// <summary>
+/// This class runs all the tests in <see cref="AbstractLspMiscellaneousFilesWorkspaceTests"/> against the base implementation.
+/// </summary>
+public sealed partial class LspMiscellaneousFilesWorkspaceProviderTests : AbstractLspMiscellaneousFilesWorkspaceTests
 {
-    public AbstractLspMiscellaneousFilesWorkspaceTests(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
+    public LspMiscellaneousFilesWorkspaceProviderTests(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
     {
     }
 
@@ -130,7 +127,7 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
         Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
 
         // Open a file that is part of a registered workspace and verify it is not present in the misc workspace.
-        var document = await AddDocumentAsync(testLspServer, TestHelpers.CreateAbsolutePath("SomeFile.cs"), markup).ConfigureAwait(false);
+        var document = await AddDocumentAsync(testLspServer, TestHelpers.CreateAbsolutePath("SomeFile.cs")).ConfigureAwait(false);
         var fileInWorkspaceUri = document.GetURI();
         await testLspServer.OpenDocumentAsync(fileInWorkspaceUri, markup).ConfigureAwait(false);
         Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
@@ -169,7 +166,7 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
         var documentPath = looseFileUri.GetDocumentFilePathFromUri();
 
         // Update the workspace to contain the loose file.
-        await AddDocumentAsync(testLspServer, documentPath, source);
+        await AddDocumentAsync(testLspServer, documentPath);
         Assert.Contains(documentPath, GetHostWorkspace(testLspServer).CurrentSolution.Projects.Single().Documents.Select(d => d.FilePath));
 
         // Verify that the manager returns the file that has been added to the main workspace.
@@ -178,34 +175,6 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
         // Make sure doc was removed from misc workspace.
         Assert.False(miscWorkspace.CurrentSolution.ContainsDocument(miscDocument.Id));
         Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
-    }
-
-    [Theory, CombinatorialData]
-    public async Task TestLooseFile_RazorFile(bool mutatingLspWorkspace)
-    {
-        var composition = Composition.AddParts(typeof(TestRazorMiscellaneousProjectInfoService));
-
-        // Create a server that supports LSP misc files and verify no misc files present.
-        await using var testLspServer = await CreateTestLspServerAsync(string.Empty, mutatingLspWorkspace, new InitializationOptions { ServerKind = WellKnownLspServerKinds.CSharpVisualBasicLspServer }, composition);
-        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
-        Assert.Null(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
-
-        // Open an empty loose file and make a request to verify it gets added to the misc workspace.
-        var looseFileUri = CreateAbsoluteDocumentUri("SomeFile.razor");
-        await testLspServer.OpenDocumentAsync(looseFileUri, "<div></div>").ConfigureAwait(false);
-
-        // Trigger a request and assert we got a file in the misc workspace.
-        await AssertFileInMiscWorkspaceAsync(testLspServer, looseFileUri).ConfigureAwait(false);
-        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
-        Assert.NotNull(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
-
-        // Trigger another request and assert we got a file in the misc workspace.
-        await AssertFileInMiscWorkspaceAsync(testLspServer, looseFileUri).ConfigureAwait(false);
-        Assert.NotNull(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
-
-        await testLspServer.CloseDocumentAsync(looseFileUri).ConfigureAwait(false);
-        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
-        Assert.Null(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
     }
 
     [Theory, CombinatorialData]
@@ -234,6 +203,65 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
 
         await testLspServer.CloseDocumentAsync(looseFileUri).ConfigureAwait(false);
         Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
+    }
+
+    [Theory, CombinatorialData]
+    public async Task TestLspTransfersFromMiscellaneousFilesToHostWorkspaceAsync(bool mutatingLspWorkspace, bool waitForWorkspace, bool fileBasedProgramContent)
+    {
+        var markup = fileBasedProgramContent ? "Console.WriteLine();" : "class C { }";
+
+        // Create a server that includes the LSP misc files workspace so we can test transfers to and from it.
+        await using var testLspServer = await CreateTestLspServerAsync(markup, mutatingLspWorkspace, new InitializationOptions { ServerKind = WellKnownLspServerKinds.CSharpVisualBasicLspServer });
+
+        // Include some Unicode characters to test URL handling.
+        using var tempRoot = new TempRoot();
+        var newDocumentFilePath = Path.Combine(tempRoot.CreateDirectory().Path, "ue25b\ud86d\udeac.cs");
+
+        // If this is file based, we're going to be inspecting the actual content on disk as a part of a dotnet run-api invocation
+        if (fileBasedProgramContent)
+            File.WriteAllText(newDocumentFilePath, markup);
+        var newDocumentUri = ProtocolConversions.CreateAbsoluteDocumentUri(newDocumentFilePath);
+
+        // Open the document via LSP before the workspace sees it.
+        await testLspServer.OpenDocumentAsync(newDocumentUri, "LSP text");
+
+        // Verify it is a miscellaneous document of some kind
+        var (_, miscDocument) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
+        Assert.NotNull(miscDocument);
+        Assert.True(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(miscDocument));
+        Assert.Equal("LSP text", (await miscDocument.GetTextAsync(CancellationToken.None)).ToString());
+
+        if (waitForWorkspace)
+        {
+            // Optionally wait for the workspace so we can test what happens if we're seeing if it's a file-based program; otherwise we can test what
+            // happens if that analysis is still happening while we're loading real solutions.
+            await testLspServer.TestWorkspace.GetService<AsynchronousOperationListenerProvider>().GetWaiter(FeatureAttribute.Workspace).ExpeditedWaitAsync();
+        }
+
+        // Make a change and verify the misc document is updated.
+        await testLspServer.InsertTextAsync(newDocumentUri, (0, 0, "More LSP text"));
+        (_, miscDocument) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
+        Assert.NotNull(miscDocument);
+        var miscText = await miscDocument.GetTextAsync(CancellationToken.None);
+        Assert.Equal("More LSP textLSP text", miscText.ToString());
+        Assert.True(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(miscDocument));
+
+        // Update the registered workspace with the new document.
+        var newDocumentId = (await AddDocumentAsync(testLspServer, newDocumentFilePath)).Id;
+
+        // Verify that the newly added document in the registered workspace is returned.
+        var (documentWorkspace, document) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
+        Assert.NotNull(document);
+        Assert.Equal(GetHostWorkspace(testLspServer), documentWorkspace);
+        Assert.False(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(document));
+        Assert.Equal(newDocumentId, document.Id);
+        // Verify we still are using the tracked LSP text for the document.
+        var documentText = await document.GetTextAsync(CancellationToken.None);
+        Assert.Equal("More LSP textLSP text", documentText.ToString());
+
+        // There should not be any other misc document in the solution anymore.
+        var matchingDocuments = await document.Project.Solution.GetTextDocumentsAsync(newDocumentUri, CancellationToken.None);
+        Assert.Single(matchingDocuments);
     }
 
     [Theory, CombinatorialData]
@@ -285,7 +313,7 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
             """, languageId: "csharp").ConfigureAwait(false);
         // Make an immediate followup request as soon as we queue the didOpen.
         // This should succeed and use the language from the didOpen.
-        var result = await testLspServer.ExecuteRequestAsync<LSP.TextDocumentPositionParams, LSP.Location[]>(LSP.Methods.TextDocumentDefinitionName,
+        var result = await testLspServer.ExecuteRequestAsync<TextDocumentPositionParams, LSP.Location[]>(Methods.TextDocumentDefinitionName,
             CreateTextDocumentPositionParams(new LSP.Location
             {
                 DocumentUri = looseFileUri,
@@ -310,124 +338,88 @@ public abstract class AbstractLspMiscellaneousFilesWorkspaceTests : AbstractLang
     }
 
     [Theory, CombinatorialData]
-    public async Task TestLspTransfersFromMiscellaneousFilesToHostWorkspaceAsync(bool mutatingLspWorkspace, bool waitForWorkspace, bool fileBasedProgramContent)
+    public async Task TestLooseFile_RazorFile(bool mutatingLspWorkspace)
     {
-        var markup = fileBasedProgramContent ? "Console.WriteLine();" : "class C { }";
+        // Create a server that supports LSP misc files and verify no misc files present.
+        await using var testLspServer = await CreateTestLspServerAsync(string.Empty, mutatingLspWorkspace, new InitializationOptions { ServerKind = WellKnownLspServerKinds.CSharpVisualBasicLspServer });
+        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
+        Assert.Null(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
 
-        // Create a server that includes the LSP misc files workspace so we can test transfers to and from it.
-        await using var testLspServer = await CreateTestLspServerAsync(markup, mutatingLspWorkspace, new InitializationOptions { ServerKind = WellKnownLspServerKinds.CSharpVisualBasicLspServer });
+        // Open an empty loose file and make a request to verify it gets added to the misc workspace.
+        var looseFileUri = CreateAbsoluteDocumentUri("SomeFile.razor");
+        await testLspServer.OpenDocumentAsync(looseFileUri, "<div></div>").ConfigureAwait(false);
 
-        // Include some Unicode characters to test URL handling.
-        using var tempRoot = new TempRoot();
-        var newDocumentFilePath = Path.Combine(tempRoot.CreateDirectory().Path, "ue25b\ud86d\udeac.cs");
+        // Trigger a request and assert we got a file in the misc workspace.
+        await AssertFileInMiscWorkspaceAsync(testLspServer, looseFileUri).ConfigureAwait(false);
+        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
+        Assert.NotNull(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
 
-        // If this is file based, we're going to be inspecting the actual content on disk as a part of a dotnet run-api invocation
-        if (fileBasedProgramContent)
-            File.WriteAllText(newDocumentFilePath, markup);
-        var newDocumentUri = ProtocolConversions.CreateAbsoluteDocumentUri(newDocumentFilePath);
+        // Trigger another request and assert we got a file in the misc workspace.
+        await AssertFileInMiscWorkspaceAsync(testLspServer, looseFileUri).ConfigureAwait(false);
+        Assert.NotNull(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
 
-        // Open the document via LSP before the workspace sees it.
-        await testLspServer.OpenDocumentAsync(newDocumentUri, "LSP text");
-
-        // Verify it is a miscellaneous document of some kind
-        var (_, miscDocument) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
-        Assert.NotNull(miscDocument);
-        Assert.True(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(miscDocument));
-        Assert.Equal("LSP text", (await miscDocument.GetTextAsync(CancellationToken.None)).ToString());
-
-        if (waitForWorkspace)
-        {
-            // Optionally wait for the workspace so we can test what happens if we're seeing if it's a file-based program; otherwise we can test what
-            // happens if that analysis is still happening while we're loading real solutions.
-            await testLspServer.TestWorkspace.GetService<AsynchronousOperationListenerProvider>().GetWaiter(FeatureAttribute.Workspace).ExpeditedWaitAsync();
-        }
-
-        // Make a change and verify the misc document is updated.
-        await testLspServer.InsertTextAsync(newDocumentUri, (0, 0, "More LSP text"));
-        (_, miscDocument) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
-        Assert.NotNull(miscDocument);
-        var miscText = await miscDocument.GetTextAsync(CancellationToken.None);
-        Assert.Equal("More LSP textLSP text", miscText.ToString());
-        Assert.True(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(miscDocument));
-
-        // Update the registered workspace with the new document.
-        var newDocumentId = (await AddDocumentAsync(testLspServer, newDocumentFilePath, "New Doc")).Id;
-
-        // Verify that the newly added document in the registered workspace is returned.
-        var (documentWorkspace, document) = await GetLspWorkspaceAndDocumentAsync(newDocumentUri, testLspServer).ConfigureAwait(false);
-        Assert.NotNull(document);
-        Assert.Equal(GetHostWorkspace(testLspServer), documentWorkspace);
-        Assert.False(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(document));
-        Assert.Equal(newDocumentId, document.Id);
-        // Verify we still are using the tracked LSP text for the document.
-        var documentText = await document.GetTextAsync(CancellationToken.None);
-        Assert.Equal("More LSP textLSP text", documentText.ToString());
-
-        // There should not be any other misc document in the solution anymore.
-        var matchingDocuments = await document.Project.Solution.GetTextDocumentsAsync(newDocumentUri, CancellationToken.None);
-        Assert.Single(matchingDocuments);
+        await testLspServer.CloseDocumentAsync(looseFileUri).ConfigureAwait(false);
+        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
+        Assert.Null(await GetMiscellaneousAdditionalDocumentAsync(testLspServer));
     }
 
-    private protected abstract ValueTask<Document> AddDocumentAsync(TestLspServer testLspServer, string filePath, string content);
-    private protected abstract Workspace GetHostWorkspace(TestLspServer testLspServer);
-
-    private protected static async Task<(Workspace? workspace, Document? document)> GetLspWorkspaceAndDocumentAsync(DocumentUri uri, TestLspServer testLspServer)
+    [Theory, CombinatorialData]
+    public async Task TestLooseFilesInCanonicalProject(bool mutatingLspWorkspace)
     {
-        var (workspace, _, document) = await testLspServer.GetManager().GetLspDocumentInfoAsync(CreateTextDocumentIdentifier(uri), CancellationToken.None).ConfigureAwait(false);
-        return (workspace, document as Document);
-    }
+        // Create a server that supports LSP misc files and verify no misc files present.
+        await using var testLspServer = await CreateTestLspServerAsync(string.Empty, mutatingLspWorkspace, new InitializationOptions { ServerKind = WellKnownLspServerKinds.CSharpVisualBasicLspServer });
+        Assert.Null(await GetMiscellaneousDocumentAsync(testLspServer));
 
-    private protected static async Task<(Workspace workspace, Document document)> GetRequiredLspWorkspaceAndDocumentAsync(DocumentUri uri, TestLspServer testLspServer)
-    {
-        var (workspace, document) = await GetLspWorkspaceAndDocumentAsync(uri, testLspServer);
-        Assert.NotNull(workspace);
-        Assert.NotNull(document);
-        return (workspace, document);
-    }
+        var looseFileUriOne = CreateAbsoluteDocumentUri("SomeFile.cs");
+        await testLspServer.OpenDocumentAsync(looseFileUriOne, """
+            class A
+            {
+                void M()
+                {
+                }
+            }
+            """).ConfigureAwait(false);
 
-    private protected static async ValueTask<Document?> GetMiscellaneousDocumentAsync(TestLspServer testLspServer)
-    {
-        var documents = await testLspServer.GetManagerAccessor().GetMiscellaneousDocumentsAsync(static p => p.Documents).ToImmutableArrayAsync(CancellationToken.None);
-        return documents.SingleOrDefault();
-    }
+        // Document should be initially found in a primordial misc files project
+        var (_, looseDocumentOne) = await GetLspWorkspaceAndDocumentAsync(looseFileUriOne, testLspServer).ConfigureAwait(false);
+        Assert.NotNull(looseDocumentOne);
+        Assert.Equal(1, looseDocumentOne.Project.Documents.Count());
+        Assert.Empty(looseDocumentOne.Project.MetadataReferences);
 
-    private protected static async ValueTask<TextDocument?> GetMiscellaneousAdditionalDocumentAsync(TestLspServer testLspServer)
-    {
-        var documents = await testLspServer.GetManagerAccessor().GetMiscellaneousDocumentsAsync(static p => p.AdditionalDocuments).ToImmutableArrayAsync(CancellationToken.None);
-        return documents.SingleOrDefault();
-    }
+        // Wait for the canonical project to finish loading.
+        await testLspServer.TestWorkspace.GetService<AsynchronousOperationListenerProvider>().GetWaiter(FeatureAttribute.Workspace).ExpeditedWaitAsync();
 
-    private protected static async Task AssertFileInMiscWorkspaceAsync(TestLspServer testLspServer, DocumentUri fileUri)
-    {
-        var (_, _, document) = await testLspServer.GetManager().GetLspDocumentInfoAsync(new LSP.TextDocumentIdentifier { DocumentUri = fileUri }, CancellationToken.None);
-        Assert.NotNull(document);
-        Assert.True(await testLspServer.GetManagerAccessor().IsMiscellaneousFilesDocumentAsync(document));
-    }
+        // Verify the document is found in a forked canonical project.
+        var (_, canonicalDocumentOne) = await GetLspWorkspaceAndDocumentAsync(looseFileUriOne, testLspServer).ConfigureAwait(false);
+        Assert.NotNull(canonicalDocumentOne);
+        Assert.NotEqual(looseDocumentOne, canonicalDocumentOne);
+        // Should have the appropriate generated files now that we ran a design time build
+        Assert.Contains(canonicalDocumentOne.Project.Documents, d => d.Name == "Canonical.AssemblyInfo.cs");
 
-    private async Task AssertFileInMainWorkspaceAsync(TestLspServer testLspServer, DocumentUri fileUri)
-    {
-        var (lspWorkspace, _, _) = await testLspServer.GetManager().GetLspDocumentInfoAsync(new LSP.TextDocumentIdentifier { DocumentUri = fileUri }, CancellationToken.None).ConfigureAwait(false);
-        Assert.Equal(GetHostWorkspace(testLspServer), lspWorkspace);
-    }
+        // Add another loose virtual document and verify it goes into a forked canonical project.
+        var looseFileUriTwo = ProtocolConversions.CreateAbsoluteDocumentUri(@"vscode-notebook-cell://dev-container/test.cs");
+        await testLspServer.OpenDocumentAsync(looseFileUriTwo, """
+            class Other
+            {
+                void OtherMethod()
+                {
+                }
+            }
+            """).ConfigureAwait(false);
 
-    private static async Task<LSP.Hover> RunGetHoverAsync(TestLspServer testLspServer, LSP.Location caret)
-    {
-        var result = await testLspServer.ExecuteRequestAsync<LSP.TextDocumentPositionParams, LSP.Hover>(LSP.Methods.TextDocumentHoverName,
-            CreateTextDocumentPositionParams(caret), CancellationToken.None);
-        Contract.ThrowIfNull(result);
-        return result;
-    }
+        var (_, canonicalDocumentTwo) = await GetLspWorkspaceAndDocumentAsync(looseFileUriTwo, testLspServer).ConfigureAwait(false);
 
-    // This is a test version of the real service which lives in the Razor EA, which is not referenced here
-    [PartNotDiscoverable]
-    [ExportLanguageService(typeof(IMiscellaneousProjectInfoService), "Razor"), Shared]
-    [method: ImportingConstructor]
-    [method: Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-    private sealed class TestRazorMiscellaneousProjectInfoService() : IMiscellaneousProjectInfoService
-    {
-        public string ProjectLanguageOverride => LanguageNames.CSharp;
-        public bool AddAsAdditionalDocument => true;
+        // Wait for the canonical project to finish loading.
+        await testLspServer.TestWorkspace.GetService<AsynchronousOperationListenerProvider>().GetWaiter(FeatureAttribute.Workspace).ExpeditedWaitAsync();
+        (_, canonicalDocumentTwo) = await GetLspWorkspaceAndDocumentAsync(looseFileUriTwo, testLspServer).ConfigureAwait(false);
 
-        public IEnumerable<AnalyzerReference>? GetAnalyzerReferences(SolutionServices services) => null;
+        Assert.NotNull(canonicalDocumentTwo);
+        Assert.NotEqual(canonicalDocumentOne.Project.Id, canonicalDocumentTwo.Project.Id);
+        Assert.DoesNotContain(canonicalDocumentTwo.Project.Documents, d => d.Name == looseDocumentOne.Name);
+        // Semantic diagnostics are not expected due to absence of top-level statements
+        Assert.False(canonicalDocumentTwo.Project.State.HasAllInformation);
+        // Should have the appropriate generated files from the base misc files project now that we ran a design time build
+        Assert.Contains(canonicalDocumentTwo.Project.Documents, d => d.Name == "Canonical.AssemblyInfo.cs");
     }
 }
+
